+++ packages/core-js/stage/0.js
@@ -2,5 +2,6 @@ require('../proposals/asap');
 require('../proposals/efficient-64-bit-arithmetic');
 require('../proposals/string-at');
 require('../proposals/pattern-matching');
+require('../proposals/promise-all-settled');
 
 module.exports = require('./1');